remove anchor-nd and move serialization/deserialization to vec crate
