--- anchor-nd/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "anchor-nd"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anchor-lang = { version = "0.29.0", optional = true}
-vectors = { path = "../vectors" }
-cvt = { path = "../cvt" }
-
-[features]
-include-anchor = ["dep:anchor-lang"]
