+++ vectors/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anchor-lang = { version = "0.29.0"}
+borsh = { version = "0.10" }
 cvt = { path = "../cvt" }
 nondet = { path = "../nondet" }
